adds a unit test
